connects Pad component
