+++ package.json
@@ -17,8 +17,10 @@
     "react-dom": "^15.5.4",
     "react-redux": "^5.0.4",
     "redux": "^3.6.0",
+    "redux-devtools-extension": "^2.13.2",
     "rxjs": "^5.3.1",
-    "web-animations-js": "^2.2.5"
+    "web-animations-js": "^2.2.5",
+    "prop-types": "^15.5.10"
   },
   "devDependencies": {
     "directory-named-webpack-plugin": "^2.1.0",
