+++ src/Board/Board.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import styles from './Board.css';
 
 import {COLORS} from 'src/utils/constants'
@@ -7,29 +6,12 @@ import {COLORS} from 'src/utils/constants'
 import Console from 'src/Console'
 import Pad from 'src/Pad'
 
-Board.propTypes = {
-  strictMode: PropTypes.bool.isRequired,
-  onStrictClick: PropTypes.func.isRequired,
-  onResetClick: PropTypes.func.isRequired,
-  onPadClick: PropTypes.func.isRequired,
-  observable: PropTypes.object
-}
-
-export default function Board (props) {
+export default function Board () {
   return (
     <div className={styles.board}>
-      <Console
-        strictMode={props.strictMode}
-        onStrictClick={props.onStrictClick}
-        onResetClick={props.onResetClick}
-      />
+      <Console />
       {COLORS.map((color) =>
-        <Pad
-          key={color}
-          color={color}
-          onClick={props.onPadClick}
-          observable={props.observable}
-        />
+        <Pad color={color} key={color} />
       )}
     </div>
   )
